cryptotrace/cryptotrace-ocr/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.6.1
python-jose==3.3.0
Pillow==10.2.0
pytesseract==0.3.10
httpx==0.26.0
demjson3==3.0.6 